requirements.txt
# LLM
litellm>=1.40.0
httpx>=0.27.0

# Search APIs
arxiv>=2.1.0
semanticscholar>=0.8.0

# PDF
PyMuPDF>=1.24.0

# Web crawling
playwright>=1.44.0

# LaTeX parsing
pylatexenc>=2.10

# Config & validation
pydantic>=2.7.0
pydantic-settings>=2.3.0
pyyaml>=6.0.1

# CLI & display
click>=8.1.7
rich>=13.7.0

# Utilities
python-slugify>=8.0.4
tenacity>=8.3.0
jinja2>=3.1.4
eval-type-backport>=0.2.0  # Python 3.9 compat for PEP 604 type unions

# FastAPI + server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis[hiredis]>=5.0.0

# Env loading
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
